src/test/test_advanced_features.py
#!/usr/bin/env python3
"""
Advanced Features Test Suite.

Tests for Celestra advanced features:
- DependencyManager (dependency management)
- WaitCondition (wait conditions)
- CustomResource (custom resources and operators)
- CostOptimization (cost management and optimization)
"""

import pytest
import yaml

from src.celestra import (
    DependencyManager, WaitCondition, CustomResource, CostOptimization, 
    DeploymentStrategy, ExternalServices, App, StatefulApp
)
from .utils import TestHelper, AssertionHelper, MockKubernetesCluster, assert_valid_kubernetes_resource


class TestDependencyManager:
    """Test cases for the DependencyManager class (dependency management)."""
    
    def setup_method(self):
        """Set up test environment."""
        self.helper = TestHelper("dependency_manager_tests")
        self.cluster = MockKubernetesCluster()
    
    def teardown_method(self):
        """Clean up test environment."""
        self.helper.cleanup()
    
    def test_basic_dependency_management(self):
        dep_manager = (DependencyManager("test-deps")
                       .add_service_dependency("db-dep", "user-service"))
        
        assert dep_manager._name == "test-deps"
        assert len(dep_manager._dependencies) == 1

    def test_wait_conditions(self):
        dep_manager = (DependencyManager("wait-deps")
                       .add_database_dependency("postgres-dep", "postgres"))
        
        assert dep_manager._name == "wait-deps"
        assert len(dep_manager._dependencies) == 1

    def test_external_service_dependencies(self):
        dep_manager = (DependencyManager("external-deps")
                       .add_api_dependency("payment-dep", "https://api.payment.com"))
        
        assert dep_manager._name == "external-deps"
        assert len(dep_manager._dependencies) == 1

    def test_multi_service_orchestration(self):
        dep_manager = (DependencyManager("multi-deps")
                       .add_service_dependency("dep-a", "service-a")
                       .add_service_dependency("dep-b", "service-b"))
        
        assert len(dep_manager._dependencies) == 2

    def test_health_check_based_dependencies(self):
        dep_manager = (DependencyManager("health-deps")
                       .add_service_dependency("api-dep", "api-service"))
        
        assert len(dep_manager._dependencies) == 1

    def test_custom_dependency_checks(self):
        dep_manager = DependencyManager("custom-deps")
        assert dep_manager._name == "custom-deps"

    def test_conditional_dependencies(self):
        dep_manager = DependencyManager("conditional-deps")
        assert dep_manager._name == "conditional-deps"

    def test_dependency_groups(self):
        dep_manager = DependencyManager("group-deps")
        assert dep_manager._name == "group-deps"


class TestWaitCondition:
    def test_basic_wait_condition(self):
        wait_condition = WaitCondition("test-wait")
        assert wait_condition._name == "test-wait"

    def test_service_ready_condition(self):
        wait_condition = WaitCondition("service-wait")
        assert wait_condition._name == "service-wait"

    def test_pod_ready_condition(self):
        wait_condition = WaitCondition("pod-wait")
        assert wait_condition._name == "pod-wait"

    def test_custom_condition(self):
        wait_condition = WaitCondition("custom-wait")
        assert wait_condition._name == "custom-wait"

    def test_timeout_conditions(self):
        wait_condition = WaitCondition("timeout-wait")
        assert wait_condition._name == "timeout-wait"

    def test_multiple_conditions(self):
        wait_condition = WaitCondition("multi-wait")
        assert wait_condition._name == "multi-wait"


class TestDeploymentStrategy:
    def test_rolling_update_strategy(self):
        strategy = DeploymentStrategy("rolling-strategy")
        assert strategy._name == "rolling-strategy"

    def test_blue_green_strategy(self):
        strategy = DeploymentStrategy("blue-green-strategy")
        assert strategy._name == "blue-green-strategy"

    def test_canary_strategy(self):
        strategy = DeploymentStrategy("canary-strategy")
        assert strategy._name == "canary-strategy"

    def test_recreate_strategy(self):
        strategy = DeploymentStrategy("recreate-strategy")
        assert strategy._name == "recreate-strategy"

    def test_custom_strategy(self):
        strategy = DeploymentStrategy("custom-strategy")
        assert strategy._name == "custom-strategy"


class TestExternalServices:
    def test_database_integration(self):
        external = ExternalServices("db-external")
        assert external._name == "db-external"

    def test_api_gateway_integration(self):
        external = ExternalServices("api-external")
        assert external._name == "api-external"

    def test_message_queue_integration(self):
        external = ExternalServices("mq-external")
        assert external._name == "mq-external"

    def test_storage_service_integration(self):
        external = ExternalServices("storage-external")
        assert external._name == "storage-external"

    def test_monitoring_service_integration(self):
        external = ExternalServices("monitoring-external")
        assert external._name == "monitoring-external"


class TestCostOptimization:
    def test_resource_right_sizing(self):
        optimizer = CostOptimization("resource-optimizer")
        assert optimizer._name == "resource-optimizer"

    def test_instance_scheduling(self):
        optimizer = CostOptimization("schedule-optimizer")
        assert optimizer._name == "schedule-optimizer"

    def test_storage_optimization(self):
        optimizer = CostOptimization("storage-optimizer")
        assert optimizer._name == "storage-optimizer"

    def test_network_cost_optimization(self):
        optimizer = CostOptimization("network-optimizer")
        assert optimizer._name == "network-optimizer"

    def test_cost_monitoring(self):
        optimizer = CostOptimization("cost-monitor")
        assert optimizer._name == "cost-monitor"


class TestCustomResource:
    def test_basic_custom_resource(self):
        custom_resource = CustomResource("test-crd")
        assert custom_resource._name == "test-crd"

    def test_crd_with_validation(self):
        custom_resource = CustomResource("validated-crd")
        assert custom_resource._name == "validated-crd"

    def test_crd_with_subresources(self):
        custom_resource = CustomResource("subresource-crd")
        assert custom_resource._name == "subresource-crd"

    def test_crd_with_additional_printer_columns(self):
        custom_resource = CustomResource("printer-crd")
        assert custom_resource._name == "printer-crd"

    def test_crd_with_webhook(self):
        custom_resource = CustomResource("webhook-crd")
        assert custom_resource._name == "webhook-crd"


if __name__ == "__main__":
    pytest.main([__file__, "-v"]) 