src/requirements.txt
# Celestra DSL Requirements

# Core Dependencies
PyYAML>=6.0                # YAML processing (used throughout DSL)
pytest>=7.4.0              # Testing framework
pytest-cov>=4.1.0          # Test coverage
pytest-json-report>=1.5.0  # JSON test reporting (for test runner)

# Development Dependencies (optional but recommended)
black>=23.7.0              # Code formatting
mypy>=1.5.0                # Type checking
flake8>=6.0.0              # Linting

# Documentation Dependencies
mkdocs>=1.5.0              # Documentation generator
mkdocs-material>=9.1.0     # Material theme for MkDocs
mkdocs-awesome-pages-plugin>=2.9.0  # Better page organization
mkdocstrings[python]>=0.22.0        # Python API documentation
pymdown-extensions>=10.0   # Additional markdown extensions
mkdocs-mermaid2-plugin>=1.1.0       # Mermaid diagrams support

# Optional Runtime Dependencies (auto-detected)
# pathlib - built into Python 3.4+
# typing - built into Python 3.5+
# abc - built into Python standard library
